modules/grpc-kafka-producer/requirements.txt
kafka-python==2.1.4
grpcio==1.62.3
grpcio-tools==1.62.3